SM-61 Logout and styles
