+++ public/locales/en/navAndFooter.json
@@ -5,5 +5,6 @@
   "contact": "contact",
   "login": "log in",
   "cart": "your cart",
-  "contactData": "Contact data"
+  "contactData": "Contact data",
+  "toastOut": "You are log out"
 }
