+++ public/locales/pl/navAndFooter.json
@@ -5,5 +5,6 @@
   "contact": "kontakt",
   "login": "zaloguj",
   "cart": "twój koszyk",
-  "contactData": "Dane kontaktowe"
+  "contactData": "Dane kontaktowe",
+  "toastOut": "Zostałeś wylogowany"
 }
