+++ src/features/layout/nav/Nav.tsx
@@ -1,7 +1,9 @@
+import toast from 'react-hot-toast';
 import { AiOutlineClose, AiOutlineMenu } from 'react-icons/ai';
 import { BiUser } from 'react-icons/bi';
 import { BsCart3 } from 'react-icons/bs';
 import { NavLink } from 'react-router-dom';
+import { useDispatch, useSelector } from 'react-redux';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -9,10 +11,18 @@ import Backdrop from './Backdrop';
 import LanguageButtons from './LanguageButtons';
 import NavBar from './NavBar';
 import logo from '../../../assets/s-logo.svg';
+import { RootState } from '../../../app/store';
 import { activeLogCart, navBurgerStyles, navStyles } from './navStyles';
+import { authActions } from '../../loginPage/authSlice';
+import { userActions } from '../../loginPage/userSlice';
 
 const Nav = () => {
   const { t, i18n } = useTranslation('navAndFooter');
+  const dispatch = useDispatch();
+
+  const isAuth = useSelector(
+    (state: RootState) => state.authentication.isAuthenticated,
+  );
 
   const changeLanguage = (lng: string) => {
     i18n.changeLanguage(lng);
@@ -22,6 +32,19 @@ const Nav = () => {
   const handleNav = () => {
     setNav(!nav);
   };
+
+  const logoutHandler = () => {
+    const toastId = toast.loading('Loading...');
+
+    dispatch(authActions.logout());
+    dispatch(userActions.userLogout());
+
+    toast.success(t('toastOut'), {
+      id: toastId,
+    });
+    return;
+  };
+
   return (
     <div className="mx-auto flex h-20 w-full max-w-7xl items-center justify-between sm:h-36 sm:px-3 md:h-44 md:px-6 lg:px-8">
       <div
@@ -35,6 +58,15 @@ const Nav = () => {
       </div>
       <section className="hidden sm:block">
         <div className="flex justify-end gap-8 p-4 pl-8 text-2xl">
+          {isAuth && (
+            <button
+              onClick={logoutHandler}
+              tabIndex={0}
+              className="border-1 rounded border border-black p-1 px-2 text-xs"
+            >
+              Wyloguj
+            </button>
+          )}
           <NavLink
             to="/login"
             aria-label="login"
